navigation being set up
