+++ App.js
@@ -1,130 +1,97 @@
-import 'react-native-gesture-handler';
+import * as React from 'react';
+import { Button, View, Text } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import { StatusBar } from 'expo-status-bar';
-import React from 'react';
-import { StyleSheet, Platform, Text, View, Image, TouchableOpacity } from 'react-native';
-import * as ImagePicker from 'expo-image-picker';
-import * as Sharing from 'expo-sharing'; 
-import uploadToAnonymousFilesAsync from 'anonymous-files'; 
-import reading from './assets/reading.png';
-import sharePhoto from './assets/sharePhoto.png';
-import largeTitle from './assets/largeTitle.png';
+import { createStackNavigator } from '@react-navigation/stack';
 
-export default function App() {
-  const [selectedImage, setSelectedImage] = React.useState(null);
+function HomeScreen({navigation}) {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Home Screen</Text>
+      <Button
+        title="Go to Favorites"
+        onPress={() => navigation.navigate('Favorites')}
+      />
+      <Button
+        title="Go to Shop"
+        onPress={() => navigation.navigate('Shop')}
+      />
+    </View>
+  );
+}
 
-  let openImagePickerAsync = async () => {
-    let permissionResult = await ImagePicker.requestCameraRollPermissionsAsync();
+function FavoritesScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Favorites Screen</Text>
+    </View>
+  )
+}
 
-    if (permissionResult.granted === false) {
-      alert("Permission to access camera roll is required!");
-      return;
-    }
+function ShopScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Shop Screen</Text>
+    </View>
+  )
+}
 
-    let pickerResult = await ImagePicker.launchImageLibraryAsync();
-    
-    if (pickerResult.cancelled === true) {
-      return;
-    }
+function VirtualCoffeeReadingScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Virtual Coffee Reading Screen</Text>
+    </View>
+  )
+}
 
-    // workaround for web sharing
-    if (Platform.OS === 'web') {
-      let remoteUri = await uploadToAnonymousFilesAsync(pickerResult.uri);
-      setSelectedImage({ localUri: pickerResult.uri, remoteUri });
-    } else {
-      setSelectedImage({ localUri: pickerResult.uri, remoteUri: null });
-    } 
-  };
+function VirtualLoadingScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Virtual Loading Screen</Text>
+    </View>
+  )
+}
 
-  let openShareDialogAsync = async () => {
-    if (!(await Sharing.isAvailableAsync())) {
-      alert(`The image is available for sharing at: ${selectedImage.remoteUri}`);
-      return;
-    }
+function PhotoReadingScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Photo Reading Screen</Text>
+    </View>
+  )
+}
 
-    await Sharing.shareAsync(selectedImage.localUri);
-  }; 
+function SignUpScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Sign Up Screen</Text>
+    </View>
+  )
+}
 
-  if(selectedImage !== null) {
-    return (
-      // must wrap all code in navigation container
-      <NavigationContainer>
-        <View style={styles.container}>
-          <Image
-            source={{ uri: selectedImage.localUri }}
-            style={styles.thumbnail}
-          />
-          <TouchableOpacity onPress={openShareDialogAsync} style={styles.button}>
-            <Text style={styles.buttonText}>Share this photo</Text>
-          </TouchableOpacity>
-        </View>
-      </NavigationContainer>
-    );
-  }
+function SignInScreen() {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Sign In Screen</Text>
+    </View>
+  )
+}
 
+const Stack = createStackNavigator();
+
+function App() {
   return (
     <NavigationContainer>
-      <View style={styles.container}>
-        <Image source={largeTitle} style={styles.appLogo} />
-        <View style={styles.logoContainer}>
-          <Image source={reading} style={styles.logo} />
-          <Image source={sharePhoto} style={styles.logo} />
-        </View>
-        <Text style={styles.instructions}>
-          To share a photo from your phone with a friend, just press the button below!
-        </Text>
-        <StatusBar style="auto" />
-
-        <TouchableOpacity
-          onPress={openImagePickerAsync}
-          style={styles.button}>
-          <Text style={styles.buttonText}>Pick a photo</Text>
-        </TouchableOpacity>
-      </View>
+      <Stack.Navigator>
+        <Stack.Screen name="Home" component={HomeScreen} />
+        <Stack.Screen name="Favorites" component={FavoritesScreen} />
+        <Stack.Screen name="Virtual" component={VirtualCoffeeReadingScreen} />
+        <Stack.Screen name="VirtualLoading" component={VirtualLoadingScreen} />
+        <Stack.Screen name="PhotoReading" component={PhotoReadingScreen} />
+        <Stack.Screen name="Shop" component={ShopScreen} />
+        <Stack.Screen name="SignUp" component={SignUpScreen} />
+        <Stack.Screen name="SignIn" component={SignInScreen} />
+      </Stack.Navigator>
     </NavigationContainer>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#E5E5E5',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  logoContainer: {
-    flexDirection: "row",
-    padding: 20,
-  },
-  logo: {
-    width: 160,
-    height: 130,
-    marginHorizontal: 10,
-  },
-  appLogo: {
-    width: 320,
-    height: 38,
-    paddingBottom: 30
-  },
-  instructions: {
-    color: '#888',
-    fontSize: 18,
-    marginHorizontal: 15,
-  }, 
-  button: {
-    backgroundColor: "blue",
-    padding: 20,
-    borderRadius: 5,
-  },
-  buttonText: {
-    fontSize: 20,
-    color: '#fff',
-  },
-  thumbnail: {
-    width: 300,
-    height: 300,
-    resizeMode: "contain"
-  },
-
-});
-
+export default App;
